redundant types and deprecations (#344)
* fix: remove redundant type declaration

* fix: remove unnecessary rand.seed declaration

* fix: handle ioutil deprecation

---------

+++ pkg/miekg/answers.go
@@ -319,12 +319,12 @@ func isDigit(b byte) bool { return b >= '0' && b <= '9' }
 
 func dddToByte(s []byte) byte {
 	_ = s[2] // bounds check hint to compiler; see golang.org/issue/14808
-	return byte((s[0]-'0')*100 + (s[1]-'0')*10 + (s[2] - '0'))
+	return (s[0]-'0')*100 + (s[1]-'0')*10 + (s[2] - '0')
 }
 
 func dddStringToByte(s string) byte {
 	_ = s[2] // bounds check hint to compiler; see golang.org/issue/14808
-	return byte((s[0]-'0')*100 + (s[1]-'0')*10 + (s[2] - '0'))
+	return (s[0]-'0')*100 + (s[1]-'0')*10 + (s[2] - '0')
 }
 
 func nextByte(s string, offset int) (byte, int) {
